rofi-mediaplayer: grab keyboard as soon as possible
Otherwise, we may miss the shortcut input. With -no-lazy-grab, rofi
grabs the keyboard early and wait for half a second before giving up.
Without, it will try for 0 second. So, if i3 still has the grab, it
will try again every millisecond for 5 seconds. However, this needs to
get scheduled after everything else, so in practice, this can take
some time. I get better results with -no-lazy-grab.

Also, we could use --release from i3. In this case, we are guaranteed
that rofi can grab the keyboard during the first try. But, after
trying a bit, I think this is still better with -no-lazy-grab: the
shell script needs to be executed, rofi needs to be executed. It's
better to do that while the user did not depress the key yet.