i3-companion: use new protocol format
It's simple enough and I don't want to spawn a polybar-msg. Maybe at a
later point, we may be smarter. The problem with the old socket is
that if you send messages fast enough on it, polybar may believe they
are one message since it is not length-encoded on its side and it's a
simple FIFO file (so no message boundaries). Dunno why I didn't get a
problem until now.