firefox: don't allow sites to override shortcuts
Each key can have three states:

- reserved
- not reserved
- unset

When reserved, it cannot be overridden by any site. When not reserved,
it can be. When unset, it depends on the permissions.default.shortcuts.
Setting it to 2 will make the key reserved, otherwise not.

You can find if a key is reserved here:
 ./browser/base/content/browser-sets.inc

The key depends on the locale. For some reason, the default one is in
./browser/locales/en-US/browser/browserSets.ftl. It seems accel means
Ctrl for Unix.